Separate admin dashboard into multiple sub-components
