+++ frontend/src/app/app.module.js
@@ -1,4 +1,7 @@
 import { AppComponent } from "./app.component.js";
 import { LoginComponent } from "./components/login/login.component.js";
-
 import { AdminDashboardComponent } from "./components/dashboard/admin-dashboard/admin-dashboard.component.js";
+import { AdminSidebarComponent } from "./components/dashboard/admin-dashboard/admin-sidebar/admin-sidebar.component.js";
+import { AdminMainComponent } from "./components/dashboard/admin-dashboard/admin-main/admin-main.component.js";
+import { StudentsMgtComponent } from "./components/dashboard/admin-dashboard/admin-main/students-management/students-mgt.component.js";
+import { SearchBarComponent } from "./components/search-bar/search-bar.component.js";
