+++ frontend/src/app/components/dashboard/admin-dashboard/admin-main/admin-main.component.js
@@ -1,0 +1,28 @@
+export class AdminMainComponent extends HTMLElement {
+    constructor() {
+        super();
+    }
+
+    connectedCallback() {
+        this.innerHTML = this.#render();
+    }
+
+    disconnectedCallback() {
+
+    }
+
+    #render() {
+        return `
+        <div class="xl:pl-72 se-right-container h-full">
+            <search-bar></search-bar>
+            <main class="" style="height: calc(100% - 4rem);">
+                <students-mgt></students-mgt>
+            </main>
+        </div>
+        `;
+    }
+
+    
+}
+
+customElements.define("admin-main", AdminMainComponent);
