+++ frontend/src/app/helpers/datetime.helper.js
@@ -1,4 +1,4 @@
- export const formatDatetime = function(datetime) {
+ export const formatDate = function(datetime) {
     //Buoc 1: Lay 10 ki tu dau cua datetime 
     const dt = datetime.slice(0,10);
